refactor(support): split the 1,153-line auto-approver file along its seams
XcodePermissionDialogAutoApprover.swift held four separable concerns:
the snapshot/decision value types, the pure matching policy, the AX
abstraction with its live client, and the polling approver itself. They
now live in PermissionDialogSnapshot/Matcher/AXClient plus the approver
file - pure file moves within the module, no API change. Also add the
ProxyStdioTransport import that StdioAdapterIntegrationTests had been
getting through the deleted re-export.
