+++ Sources/ProxyXcodeSupport/XcodePermissionDialogAutoApprover.swift
@@ -4,648 +4,6 @@ import Foundation
 import Logging
 import ProxyCore
 
-package enum XcodePermissionDialogAccessibilityStatus: Sendable {
-    case trusted
-    case untrusted
-}
-
-package struct XcodePermissionDialogButtonSnapshot: Equatable, Sendable {
-    package let title: String?
-    package let role: String?
-    package let subrole: String?
-    package let identifier: String?
-
-    package init(
-        title: String? = nil,
-        role: String? = nil,
-        subrole: String? = nil,
-        identifier: String? = nil
-    ) {
-        self.title = title
-        self.role = role
-        self.subrole = subrole
-        self.identifier = identifier
-    }
-}
-
-package struct XcodePermissionDialogWindowSnapshot: Equatable, Sendable {
-    package let processBundleIdentifier: String?
-    package let title: String
-    package let textValues: [String]
-    package let role: String?
-    package let subrole: String?
-    package let windowIdentifier: String?
-    package let isModal: Bool
-    package let isMain: Bool?
-    package let isMinimized: Bool?
-    package let document: String?
-    package let childCount: Int
-    package let hasProxy: Bool
-    package let defaultButton: XcodePermissionDialogButtonSnapshot?
-    package let cancelButton: XcodePermissionDialogButtonSnapshot?
-
-    package init(
-        processBundleIdentifier: String? = nil,
-        title: String,
-        textValues: [String],
-        role: String? = nil,
-        subrole: String? = nil,
-        windowIdentifier: String? = nil,
-        isModal: Bool,
-        isMain: Bool? = nil,
-        isMinimized: Bool? = nil,
-        document: String? = nil,
-        childCount: Int = 0,
-        hasProxy: Bool = false,
-        defaultButton: XcodePermissionDialogButtonSnapshot? = nil,
-        cancelButton: XcodePermissionDialogButtonSnapshot? = nil
-    ) {
-        self.processBundleIdentifier = processBundleIdentifier
-        self.title = title
-        self.textValues = textValues
-        self.role = role
-        self.subrole = subrole
-        self.windowIdentifier = windowIdentifier
-        self.isModal = isModal
-        self.isMain = isMain
-        self.isMinimized = isMinimized
-        self.document = document
-        self.childCount = childCount
-        self.hasProxy = hasProxy
-        self.defaultButton = defaultButton
-        self.cancelButton = cancelButton
-    }
-}
-
-package struct XcodePermissionDialogMatchDecision: Equatable, Sendable {
-    package let fingerprint: String
-    package let defaultButtonTitle: String
-
-    package init(fingerprint: String, defaultButtonTitle: String) {
-        self.fingerprint = fingerprint
-        self.defaultButtonTitle = defaultButtonTitle
-    }
-}
-
-package enum XcodePermissionDialogMatcher {
-    private static let allowedProcessBundleIdentifiers = normalizedCandidates([
-        "com.apple.dt.Xcode",
-        "com.apple.dt.ExternalViewService",
-        "com.apple.dt.Xcode.DeveloperSystemPolicyService",
-    ])
-    private static let allowedWindowRoles = normalizedCandidates([
-        "AXWindow"
-    ])
-    private static let allowedWindowSubroles = normalizedCandidates([
-        "AXDialog",
-        "AXSystemDialog",
-    ])
-
-    package static func decision(
-        for snapshot: XcodePermissionDialogWindowSnapshot,
-        processID: pid_t,
-        agentPathCandidates: Set<String> = [],
-        assistantNameCandidates: Set<String>,
-        serverProcessIDCandidates: Set<pid_t> = [ProcessInfo.processInfo.processIdentifier]
-    ) -> XcodePermissionDialogMatchDecision? {
-        guard passesStructuralChecks(snapshot) else {
-            return nil
-        }
-
-        let normalizedAgentPaths = normalizedCandidates(agentPathCandidates)
-        let normalizedAssistantNames = normalizedCandidates(assistantNameCandidates)
-        let pidCandidates = normalizedPIDCandidates(serverProcessIDCandidates)
-        let normalizedTextNodes = normalizedTextNodes(for: snapshot)
-        let defaultButtonDescription = normalizedButtonDescription(snapshot.defaultButton)
-        guard containsAssistantNameAndPID(
-            in: normalizedTextNodes,
-            agentPathCandidates: normalizedAgentPaths,
-            assistantNameCandidates: normalizedAssistantNames,
-            serverProcessIDCandidates: pidCandidates,
-            defaultButtonDescription: defaultButtonDescription
-        ) else {
-            return nil
-        }
-
-        return XcodePermissionDialogMatchDecision(
-            fingerprint: fingerprint(for: snapshot, processID: processID),
-            defaultButtonTitle: defaultButtonDescription
-        )
-    }
-
-    package static func fingerprint(
-        for snapshot: XcodePermissionDialogWindowSnapshot,
-        processID: pid_t
-    ) -> String {
-        let textFingerprint = normalizedTextNodes(for: snapshot).joined(separator: "\u{1F}")
-        let bundleFingerprint = normalizedText(snapshot.processBundleIdentifier) ?? ""
-        let roleFingerprint = normalizedText(snapshot.role) ?? ""
-        let subroleFingerprint = normalizedText(snapshot.subrole) ?? ""
-        return [
-            "\(processID)",
-            bundleFingerprint,
-            roleFingerprint,
-            subroleFingerprint,
-            "\(snapshot.isModal)",
-            textFingerprint,
-        ].joined(separator: "|")
-    }
-
-    package static func passesStructuralChecks(_ snapshot: XcodePermissionDialogWindowSnapshot) -> Bool {
-        guard
-            let normalizedBundleIdentifier = normalizedText(snapshot.processBundleIdentifier),
-            allowedProcessBundleIdentifiers.contains(normalizedBundleIdentifier)
-        else {
-            return false
-        }
-        guard snapshot.isModal else {
-            return false
-        }
-        guard snapshot.defaultButton != nil else {
-            return false
-        }
-        guard snapshot.isMinimized != true else {
-            return false
-        }
-        if let normalizedRole = normalizedText(snapshot.role),
-           allowedWindowRoles.contains(normalizedRole) == false {
-            return false
-        }
-        if let normalizedSubrole = normalizedText(snapshot.subrole),
-           allowedWindowSubroles.contains(normalizedSubrole) == false {
-            return false
-        }
-        if looksLikeNormalWorkspaceWindow(snapshot) {
-            return false
-        }
-        return true
-    }
-
-    private static func looksLikeNormalWorkspaceWindow(_ snapshot: XcodePermissionDialogWindowSnapshot) -> Bool {
-        let hasDocument = normalizedText(snapshot.document) != nil
-        return snapshot.isMain == true && (hasDocument || snapshot.hasProxy)
-    }
-
-    private static func normalizedTextNodes(for snapshot: XcodePermissionDialogWindowSnapshot) -> [String] {
-        ([snapshot.title] + snapshot.textValues).compactMap(normalizedText)
-    }
-
-    private static func containsAssistantNameAndPID(
-        in normalizedTextNodes: [String],
-        agentPathCandidates: Set<String>,
-        assistantNameCandidates: Set<String>,
-        serverProcessIDCandidates: Set<String>,
-        defaultButtonDescription: String
-    ) -> Bool {
-        let containsPath = normalizedTextNodes.contains { text in
-            agentPathCandidates.contains(where: { text.contains($0) })
-        }
-        if assistantNameCandidates.isEmpty {
-            return containsPath
-        }
-
-        let sameNodeMatch = normalizedTextNodes.contains { text in
-            serverProcessIDCandidates.contains(where: { containsNumericToken($0, in: text) })
-                && assistantNameCandidates.contains(where: { text.contains($0) })
-        }
-        if sameNodeMatch {
-            return true
-        }
-
-        let containsAssistantName = normalizedTextNodes.contains { text in
-            assistantNameCandidates.contains(where: { text.contains($0) })
-        }
-        let containsPID = normalizedTextNodes.contains { text in
-            serverProcessIDCandidates.contains(where: { containsNumericToken($0, in: text) })
-        }
-        if containsAssistantName && containsPID {
-            return true
-        }
-        let containsUnmatchedPIDReference = normalizedTextNodes.contains(where: containsPIDReference)
-        if containsAssistantName && containsUnmatchedPIDReference {
-            return false
-        }
-        if containsAssistantName && !containsUnmatchedPIDReference
-            && looksLikeAllowButton(defaultButtonDescription)
-        {
-            return true
-        }
-        guard containsPID == false, containsPath else {
-            return false
-        }
-
-        let sameNodePathMatch = normalizedTextNodes.contains { text in
-            agentPathCandidates.contains(where: { text.contains($0) })
-                && assistantNameCandidates.contains(where: { text.contains($0) })
-        }
-        if sameNodePathMatch {
-            return true
-        }
-        return containsAssistantName
-    }
-
-    private static func looksLikeAllowButton(_ description: String) -> Bool {
-        let allowDescriptions: Set<String> = [
-            "allow",
-            "許可",
-            "action-button-1",
-        ]
-        return allowDescriptions.contains(description)
-    }
-
-    private static func normalizedButtonDescription(
-        _ button: XcodePermissionDialogButtonSnapshot?
-    ) -> String {
-        normalizedText(button?.title)
-            ?? normalizedText(button?.identifier)
-            ?? normalizedText(button?.role)
-            ?? "default"
-    }
-
-    private static func normalizedCandidates(_ candidates: some Sequence<String>) -> Set<String> {
-        Set(candidates.compactMap(normalizedText))
-    }
-
-    private static func normalizedPIDCandidates(_ candidates: Set<pid_t>) -> Set<String> {
-        Set(candidates.map(String.init))
-    }
-
-    private static func containsNumericToken(_ candidate: String, in text: String) -> Bool {
-        guard candidate.isEmpty == false else {
-            return false
-        }
-
-        var currentDigits = ""
-        currentDigits.reserveCapacity(candidate.count)
-
-        for scalar in text.unicodeScalars {
-            if CharacterSet.decimalDigits.contains(scalar) {
-                currentDigits.unicodeScalars.append(scalar)
-                continue
-            }
-            if currentDigits == candidate {
-                return true
-            }
-            currentDigits.removeAll(keepingCapacity: true)
-        }
-
-        return currentDigits == candidate
-    }
-
-    private static func containsPIDReference(_ text: String) -> Bool {
-        guard text.unicodeScalars.contains(where: { scalar in
-            CharacterSet.decimalDigits.contains(scalar)
-        }) else {
-            return false
-        }
-        let labels = [
-            "pid",
-            "process id",
-            "process identifier",
-            "process-id",
-            "process_identifier",
-            "プロセス",
-            "識別子",
-        ]
-        return labels.contains { text.contains($0) }
-    }
-
-    private static func normalizedText(_ text: String?) -> String? {
-        guard let text else {
-            return nil
-        }
-        let sanitizedScalars = text.unicodeScalars.filter { scalar in
-            scalar.properties.generalCategory != .format
-        }
-        let sanitized = String(String.UnicodeScalarView(sanitizedScalars))
-        let trimmed = sanitized.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard trimmed.isEmpty == false else {
-            return nil
-        }
-        return trimmed.lowercased()
-    }
-}
-
-package protocol XcodePermissionDialogAXAccessing: Sendable {
-    func authorizationStatus(promptIfNeeded: Bool) -> XcodePermissionDialogAccessibilityStatus
-    func runningXcodeProcessIDs() -> [pid_t]
-    func openWindows(for processID: pid_t) throws -> [XcodePermissionDialogAXWindow]
-    func pressDefaultButton(in window: XcodePermissionDialogAXWindow) throws
-}
-
-package final class XcodePermissionDialogAXWindow {
-    package let processID: pid_t
-    package let snapshot: XcodePermissionDialogWindowSnapshot
-    private let defaultButton: AXUIElement
-
-    package init(
-        processID: pid_t,
-        snapshot: XcodePermissionDialogWindowSnapshot,
-        defaultButton: AXUIElement
-    ) {
-        self.processID = processID
-        self.snapshot = snapshot
-        self.defaultButton = defaultButton
-    }
-
-    fileprivate func pressDefaultButton() throws {
-        let error = AXUIElementPerformAction(defaultButton, kAXPressAction as CFString)
-        guard error == .success else {
-            throw XcodePermissionDialogAXError.performActionFailed(error)
-        }
-    }
-}
-
-package enum XcodePermissionDialogAXError: Error, CustomStringConvertible {
-    case copyAttributeFailed(attribute: String, error: AXError)
-    case performActionFailed(AXError)
-
-    package var description: String {
-        switch self {
-        case .copyAttributeFailed(let attribute, let error):
-            return "AX attribute '\(attribute)' failed: \(error.rawValue)"
-        case .performActionFailed(let error):
-            return "AX action failed: \(error.rawValue)"
-        }
-    }
-}
-
-package enum XcodePermissionDialogAXFailureClassifier {
-    private static let externalViewServiceBundleIdentifier = "com.apple.dt.ExternalViewService"
-    private static let windowsAttribute = kAXWindowsAttribute as String
-
-    package static func isBenignOpenWindowsFailure(
-        _ error: Error,
-        processBundleIdentifier: String?
-    ) -> Bool {
-        guard processBundleIdentifier == externalViewServiceBundleIdentifier else {
-            return false
-        }
-        guard let error = error as? XcodePermissionDialogAXError else {
-            return false
-        }
-        guard case .copyAttributeFailed(let attribute, let axError) = error else {
-            return false
-        }
-        return attribute == windowsAttribute && axError == .cannotComplete
-    }
-}
-
-package struct LiveXcodePermissionDialogAXClient: XcodePermissionDialogAXAccessing {
-    private let maxDescendantCount = 128
-
-    package init() {}
-
-    package func authorizationStatus(promptIfNeeded: Bool) -> XcodePermissionDialogAccessibilityStatus {
-        if promptIfNeeded {
-            let options: NSDictionary = [
-                "AXTrustedCheckOptionPrompt" as NSString: true
-            ]
-            return AXIsProcessTrustedWithOptions(options) ? .trusted : .untrusted
-        }
-        return AXIsProcessTrusted() ? .trusted : .untrusted
-    }
-
-    package func runningXcodeProcessIDs() -> [pid_t] {
-        let bundleIdentifiers: Set<String> = [
-            "com.apple.dt.Xcode",
-            "com.apple.dt.ExternalViewService",
-            "com.apple.dt.Xcode.DeveloperSystemPolicyService",
-        ]
-        var processIDs = Set(NSWorkspace.shared.runningApplications.compactMap { application -> pid_t? in
-            guard let bundleIdentifier = application.bundleIdentifier else {
-                return nil
-            }
-            guard bundleIdentifiers.contains(bundleIdentifier), application.isTerminated == false else {
-                return nil
-            }
-            return application.processIdentifier
-        })
-        processIDs.formUnion(ProcessEnumeration.processIDs(named: "Xcode"))
-        return processIDs.sorted()
-    }
-
-
-    package func openWindows(for processID: pid_t) throws -> [XcodePermissionDialogAXWindow] {
-        let app = AXUIElementCreateApplication(processID)
-        return try copyElementArray(attribute: kAXWindowsAttribute as CFString, from: app).compactMap { window in
-            try makeWindow(processID: processID, window: window)
-        }
-    }
-
-    package func pressDefaultButton(in window: XcodePermissionDialogAXWindow) throws {
-        try window.pressDefaultButton()
-    }
-
-    private func makeWindow(
-        processID: pid_t,
-        window: AXUIElement
-    ) throws -> XcodePermissionDialogAXWindow? {
-        let role = copyString(attribute: kAXRoleAttribute as CFString, from: window)
-        if let role, role != kAXWindowRole as String {
-            return nil
-        }
-
-        let subrole = copyString(attribute: kAXSubroleAttribute as CFString, from: window)
-        if let subrole,
-           subrole != kAXDialogSubrole as String,
-           subrole != "AXSystemDialog" {
-            return nil
-        }
-
-        let isModal = copyBool(attribute: kAXModalAttribute as CFString, from: window) ?? false
-        guard isModal else {
-            return nil
-        }
-
-        let isMinimized = copyBool(attribute: kAXMinimizedAttribute as CFString, from: window)
-        guard isMinimized != true else {
-            return nil
-        }
-
-        let isMain = copyBool(attribute: kAXMainAttribute as CFString, from: window)
-        let document = copyString(attribute: kAXDocumentAttribute as CFString, from: window)
-        let hasProxy = copyElement(attribute: kAXProxyAttribute as CFString, from: window) != nil
-        if isMain == true && (hasNonEmptyText(document) || hasProxy) {
-            return nil
-        }
-
-        let children = (try? copyElementArray(attribute: kAXChildrenAttribute as CFString, from: window)) ?? []
-        guard let defaultButton = copyElement(attribute: kAXDefaultButtonAttribute as CFString, from: window)
-            ?? fallbackAllowButton(in: window)
-        else {
-            return nil
-        }
-        let processBundleIdentifier = NSRunningApplication(processIdentifier: processID)?.bundleIdentifier
-
-        let snapshot = XcodePermissionDialogWindowSnapshot(
-            processBundleIdentifier: processBundleIdentifier,
-            title: copyString(attribute: kAXTitleAttribute as CFString, from: window) ?? "",
-            textValues: collectTextValues(from: window),
-            role: role,
-            subrole: subrole,
-            windowIdentifier: copyString(attribute: kAXIdentifierAttribute as CFString, from: window),
-            isModal: isModal,
-            isMain: isMain,
-            isMinimized: isMinimized,
-            document: document,
-            childCount: children.count,
-            hasProxy: hasProxy,
-            defaultButton: buttonSnapshot(from: defaultButton),
-            cancelButton: copyElement(attribute: kAXCancelButtonAttribute as CFString, from: window)
-                .flatMap(buttonSnapshot(from:))
-        )
-
-        return XcodePermissionDialogAXWindow(
-            processID: processID,
-            snapshot: snapshot,
-            defaultButton: defaultButton
-        )
-    }
-
-    private func fallbackAllowButton(in root: AXUIElement) -> AXUIElement? {
-        var queue: [AXUIElement] = [root]
-        var visited = 0
-
-        while queue.isEmpty == false, visited < maxDescendantCount {
-            let element = queue.removeFirst()
-            visited += 1
-
-            if isAllowButton(element) {
-                return element
-            }
-            let children = (try? copyElementArray(attribute: kAXChildrenAttribute as CFString, from: element)) ?? []
-            queue.append(contentsOf: children)
-        }
-
-        return nil
-    }
-
-    private func isAllowButton(_ element: AXUIElement) -> Bool {
-        let role = copyString(attribute: kAXRoleAttribute as CFString, from: element)
-        guard role == kAXButtonRole as String else { return false }
-        let title = copyString(attribute: kAXTitleAttribute as CFString, from: element)
-            .flatMap(normalizedButtonText)
-        let identifier = copyString(attribute: kAXIdentifierAttribute as CFString, from: element)
-            .flatMap(normalizedButtonText)
-        return title == "allow"
-            || title == "許可"
-            || identifier == "action-button-1"
-    }
-
-    private func normalizedButtonText(_ value: String) -> String? {
-        let trimmed = value.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard trimmed.isEmpty == false else { return nil }
-        return trimmed.lowercased()
-    }
-
-    private func hasNonEmptyText(_ value: String?) -> Bool {
-        guard let value else { return false }
-        return value.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty == false
-    }
-
-    private func collectTextValues(from root: AXUIElement) -> [String] {
-        var queue: [AXUIElement] = [root]
-        var values: [String] = []
-        var visited = 0
-
-        while queue.isEmpty == false, visited < maxDescendantCount {
-            let element = queue.removeFirst()
-            visited += 1
-
-            if let title = copyString(attribute: kAXTitleAttribute as CFString, from: element) {
-                values.append(title)
-            }
-            if let value = copyString(attribute: kAXValueAttribute as CFString, from: element) {
-                values.append(value)
-            }
-            if let description = copyString(attribute: kAXDescriptionAttribute as CFString, from: element) {
-                values.append(description)
-            }
-
-            if let children = try? copyElementArray(attribute: kAXChildrenAttribute as CFString, from: element) {
-                queue.append(contentsOf: children)
-            }
-        }
-
-        var seen = Set<String>()
-        return values.filter { value in
-            let trimmed = value.trimmingCharacters(in: .whitespacesAndNewlines)
-            guard trimmed.isEmpty == false else {
-                return false
-            }
-            return seen.insert(trimmed).inserted
-        }
-    }
-
-    private func buttonSnapshot(from element: AXUIElement) -> XcodePermissionDialogButtonSnapshot {
-        XcodePermissionDialogButtonSnapshot(
-            title: copyString(attribute: kAXTitleAttribute as CFString, from: element)
-                ?? copyString(attribute: kAXDescriptionAttribute as CFString, from: element)
-                ?? copyString(attribute: kAXValueAttribute as CFString, from: element),
-            role: copyString(attribute: kAXRoleAttribute as CFString, from: element),
-            subrole: copyString(attribute: kAXSubroleAttribute as CFString, from: element),
-            identifier: copyString(attribute: kAXIdentifierAttribute as CFString, from: element)
-        )
-    }
-
-    private func copyString(attribute: CFString, from element: AXUIElement) -> String? {
-        guard let value = try? copyAttribute(attribute: attribute, from: element) else {
-            return nil
-        }
-        return value as? String
-    }
-
-    private func copyBool(attribute: CFString, from element: AXUIElement) -> Bool? {
-        guard let value = try? copyAttribute(attribute: attribute, from: element) else {
-            return nil
-        }
-        return value as? Bool
-    }
-
-    private func copyElement(attribute: CFString, from element: AXUIElement) -> AXUIElement? {
-        guard let value = try? copyAttribute(attribute: attribute, from: element) else {
-            return nil
-        }
-        guard CFGetTypeID(value) == AXUIElementGetTypeID() else {
-            return nil
-        }
-        return unsafe unsafeDowncast(value, to: AXUIElement.self)
-    }
-
-    private func copyElementArray(attribute: CFString, from element: AXUIElement) throws -> [AXUIElement] {
-        guard let value = try copyAttribute(attribute: attribute, from: element) else {
-            return []
-        }
-        if let elements = value as? [AXUIElement] {
-            return elements
-        }
-        if let array = value as? [AnyObject] {
-            return array.compactMap { candidate in
-                guard CFGetTypeID(candidate) == AXUIElementGetTypeID() else {
-                    return nil
-                }
-                return unsafe unsafeDowncast(candidate, to: AXUIElement.self)
-            }
-        }
-        return []
-    }
-
-    private func copyAttribute(attribute: CFString, from element: AXUIElement) throws -> CFTypeRef? {
-        var value: CFTypeRef?
-        let error = unsafe AXUIElementCopyAttributeValue(element, attribute, &value)
-        switch error {
-        case .success, .noValue:
-            return value
-        default:
-            throw XcodePermissionDialogAXError.copyAttributeFailed(
-                attribute: attribute as String,
-                error: error
-            )
-        }
-    }
-}
-
 package final class XcodePermissionDialogAutoApprover: @unchecked Sendable {
     package struct Dependencies: DependencyClient {
         package var axClient: any XcodePermissionDialogAXAccessing
