+++ Tests/ProxyProcessTests/StdioAdapterIntegrationTests.swift
@@ -1,3 +1,4 @@
+import ProxyStdioTransport
 import Foundation
 import NIO
 import NIOHTTP1
